Refactor styles across components to use color constants from theme
- Updated shadow colors to use `colors.primaryBlack` instead of hardcoded values.
- Changed background colors to use `colors.primaryWhite` and `colors.secondaryBeige` for consistency.
- Modified text colors to use `colors.primaryWhite` for better readability.
- Adjusted button colors to use `colors.primaryBlue` and `colors.primaryOrange` for active states.
- Removed the `ServiceStatusScreen` component as it was no longer needed.
- Updated navigation parameters to ensure optional properties are correctly marked.
