+++ src/components/PaymentInfo/styles.ts
@@ -174,7 +174,7 @@ const styles = StyleSheet.create({
     marginVertical: 8,
     ...Platform.select({
       ios: {
-        shadowColor: '#000',
+        shadowColor: colors.primaryBlack,
         shadowOffset: { width: 0, height: 2 },
         shadowOpacity: 0.1,
         shadowRadius: 4,
