+++ src/screens/private/client/Profile/ProfileScreen.tsx
@@ -4,6 +4,7 @@ import * as FileSystem from 'expo-file-system';
 import ProfileWrapper from './ProfileWrapper';
 import { useUserStore } from '@stores/User';
 import colors from '@theme/colors';
+import { HTTP_DOMAIN } from '@config/varEnvs';
 
 interface User {
   id: number;
@@ -38,19 +39,16 @@ const UserProfileScreen: React.FC = () => {
       console.log('📤 Tipo MIME detectado:', base64Image.substring(0, 50));
       console.log('📤 Tamanho do base64:', base64Image.length);
 
-      const response = await fetch(
-        `http://localhost:3000/api/user/${userId}/avatar`,
-        {
-          method: 'POST',
-          headers: {
-            'Content-Type': 'application/json',
-          },
-          body: JSON.stringify({
-            base64Image: base64Image,
-            userId: userId,
-          }),
+      const response = await fetch(`${HTTP_DOMAIN}/api/user/${userId}/avatar`, {
+        method: 'POST',
+        headers: {
+          'Content-Type': 'application/json',
         },
-      );
+        body: JSON.stringify({
+          base64Image: base64Image,
+          userId: userId,
+        }),
+      });
 
       console.log('📤 Status da resposta:', response.status);
 
@@ -69,7 +67,7 @@ const UserProfileScreen: React.FC = () => {
       const data = await response.json();
       console.log('✅ Resposta do servidor:', data);
 
-      const newUri = `http://localhost:3000/${data.avatar_uri}`;
+      const newUri = `${HTTP_DOMAIN}/${data.avatar_uri}`;
       setAvatarUri(newUri);
       Alert.alert('Sucesso', 'Avatar atualizado com sucesso!');
       return data;
@@ -100,13 +98,10 @@ const UserProfileScreen: React.FC = () => {
 
   const removeAvatar = async () => {
     try {
-      const response = await fetch(
-        `http://localhost:3000/api/user/${userId}/avatar`,
-        {
-          method: 'DELETE',
-          headers: { 'Content-Type': 'application/json' },
-        },
-      );
+      const response = await fetch(`${HTTP_DOMAIN}/api/user/${userId}/avatar`, {
+        method: 'DELETE',
+        headers: { 'Content-Type': 'application/json' },
+      });
 
       if (!response.ok) {
         const errorText = await response.text();
@@ -158,7 +153,7 @@ const UserProfileScreen: React.FC = () => {
       return;
     }
 
-    const fullAvatarUrl = `http://localhost:3000/${apiAvatarUri}`;
+    const fullAvatarUrl = `${HTTP_DOMAIN}/${apiAvatarUri}`;
 
     if (Platform.OS === 'web') {
       setAvatarUri(fullAvatarUrl);
@@ -195,9 +190,7 @@ const UserProfileScreen: React.FC = () => {
   useEffect(() => {
     const fetchUserData = async () => {
       try {
-        const response = await fetch(
-          `http://localhost:3000/api/user/${userId}`,
-        );
+        const response = await fetch(`${HTTP_DOMAIN}/api/user/${userId}`);
         if (!response.ok) throw new Error('Erro ao buscar usuário');
 
         const data: User = await response.json();
