+++ src/screens/public/RegisterScreen/styles.ts
@@ -28,7 +28,7 @@ export const styles = StyleSheet.create({
     maxWidth: 500,
     ...Platform.select({
       ios: {
-        shadowColor: '#000',
+        shadowColor: colors.primaryBlack,
         shadowOffset: { width: 0, height: 4 },
         shadowOpacity: 0.1,
         shadowRadius: 10,
